Implemented Add to and Remove from Cart Functionality
Defined and refined the ADD_ITEM and REMOVE_ITEM actions in the CartProvider so now the total of all items with the same name/id will be displayed under a single card entry in the cart. The user can also remove items from the cart and when the total of a given item reaches 0 the card entry is removed from the cart completely.